requirements.txt
requests
pandas
streamlit
schedule
python-dotenv
matplotlib
seaborn
